+++ tests/test_logout.py
@@ -1,24 +1,27 @@
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
-from config import URL
 from locators import StellarBurgersLocators
+from data import user_credentials
 
 
-class TestStellaBurgerSwitching:
-    def test_access_personal_cabinet(self, driver, user_credentials):
+class TestStellaBurgerLogout:
+    def test_logout_successful(self, driver):
+        email, password = user_credentials()
         signin_button = driver.find_element(*StellarBurgersLocators.SIGNIN_BUTTON)
         signin_button.click()
         WebDriverWait(driver, 2).until(expected_conditions.element_to_be_clickable(StellarBurgersLocators.LOGIN_BUTTON))
         email_field = driver.find_element(*StellarBurgersLocators.EMAIL_FIELD)
-        email_field.send_keys(user_credentials['email'])
+        email_field.send_keys(email)
         password_field = driver.find_element(*StellarBurgersLocators.PASSWORD_FIELD)
-        password_field.send_keys(user_credentials['password'])
+        password_field.send_keys(password)
         login_button = driver.find_element(*StellarBurgersLocators.LOGIN_BUTTON)
         login_button.click()
-        WebDriverWait(driver, 2).until(expected_conditions.element_to_be_clickable(StellarBurgersLocators.ORDER_BUTTON))
+        WebDriverWait(driver, 2).until(
+            expected_conditions.element_to_be_clickable(StellarBurgersLocators.MYACCOUNT_BUTTON))
         myaccount_button = driver.find_element(*StellarBurgersLocators.MYACCOUNT_BUTTON)
         myaccount_button.click()
-        WebDriverWait(driver, 2).until(expected_conditions.element_to_be_clickable(StellarBurgersLocators.PROFILE))
-        logo = driver.find_element(*StellarBurgersLocators.LOGO)
-        logo.click()
-        assert driver.current_url == f'{URL}/'
+        WebDriverWait(driver, 2).until(expected_conditions.element_to_be_clickable(StellarBurgersLocators.LOGOUT))
+        logout = driver.find_element(*StellarBurgersLocators.LOGOUT)
+        logout.click()
+        WebDriverWait(driver, 2).until(expected_conditions.element_to_be_clickable(StellarBurgersLocators.LOGIN_BUTTON))
+        assert 'login' in driver.current_url
